+++ docs/html/search/classes_0.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['hardspifastinterface_9',['HardSpiFastInterface',['../classace__spi_1_1HardSpiFastInterface.html',1,'ace_spi']]],
-  ['hardspiinterface_10',['HardSpiInterface',['../classace__spi_1_1HardSpiInterface.html',1,'ace_spi']]]
+  ['hardspifastinterface_13',['HardSpiFastInterface',['../classace__spi_1_1HardSpiFastInterface.html',1,'ace_spi']]],
+  ['hardspiinterface_14',['HardSpiInterface',['../classace__spi_1_1HardSpiInterface.html',1,'ace_spi']]]
 ];