+++ docs/html/search/classes_1.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['simplespifastinterface_11',['SimpleSpiFastInterface',['../classace__spi_1_1SimpleSpiFastInterface.html',1,'ace_spi']]],
-  ['simplespiinterface_12',['SimpleSpiInterface',['../classace__spi_1_1SimpleSpiInterface.html',1,'ace_spi']]]
+  ['simplespifastinterface_15',['SimpleSpiFastInterface',['../classace__spi_1_1SimpleSpiFastInterface.html',1,'ace_spi']]],
+  ['simplespiinterface_16',['SimpleSpiInterface',['../classace__spi_1_1SimpleSpiInterface.html',1,'ace_spi']]]
 ];