+++ examples/AutoBenchmark/generate_readme.py
@@ -26,7 +26,7 @@ print(f"""\
 
 This program determines the speed of various SPI implementations.
 
-**Version**: AceSPI v0.2
+**Version**: AceSPI v0.3
 
 **DO NOT EDIT**: This file was auto-generated using `make README.md`.