+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 * Unreleased
+* 0.3 (2021-08-17)
+    * Add `beginTransaction()`, `endTransaction()`, `transfer()`,
+     and `transfer16()` for compatibility with the `SPIClass` from `<SPI.h>`.
+        * Retain `send8()` and `send16()` as convenience methods that wrap
+          the repetitive calls to the above methods.
 * 0.2 (2021-07-30)
     * Add `T_CLOCK_SPEED` template parameter to `HardSpiInterface` and
       `HardSpiFastInterface`, defaulting to 8000000 (8 MHz).